requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.12.0=py36haa95532_0
aiohttp=3.7.4=py36h2bbff1b_1
argparse=1.4.0=pypi_0
astor=0.8.1=py36haa95532_0
async-timeout=3.0.1=py36haa95532_0
attrs=21.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py36haa95532_0
brotlipy=0.7.0=py36h2bbff1b_1003
ca-certificates=2021.4.13=haa95532_1
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2020.12.5=py36haa95532_0
cffi=1.14.5=py36hcd4344a_0
chardet=3.0.4=py36haa95532_1003
click=8.0.0=pyhd3eb1b0_0
coverage=5.5=py36h2bbff1b_2
cryptography=3.4.7=py36h71e12ea_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=pypi_0
cython=0.29.23=py36hd77b12b_0
dataclasses=0.8=pypi_0
decorator=4.4.2=pypi_0
dill=0.3.3=pypi_0
fire=0.4.0=pypi_0
gast=0.2.2=py36_0
gcloud=0.17.0=pypi_0
google-auth=1.30.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=py_0
googleapis-common-protos=1.53.0=pypi_0
grpcio=1.36.1=py36hc60d5dd_1
h5py=2.10.0=py36h5e291fa_0
hdf5=1.10.4=h7ebc959_0
httplib2=0.19.1=pypi_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.10=pyhd3eb1b0_0
idna_ssl=1.1.0=py36haa95532_0
imageio=2.9.0=pypi_0
importlib-metadata=3.10.0=py36haa95532_0
intel-openmp=2021.2.0=haa95532_616
jws=0.1.3=pypi_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=pypi_0
libprotobuf=3.14.0=h23ce68f_0
markdown=3.3.4=py36haa95532_0
matplotlib=3.3.4=pypi_0
mish-cuda=0.0.3=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
msgpack=1.0.2=pypi_0
msgpack-numpy=0.4.7.1=pypi_0
multidict=5.1.0=py36h2bbff1b_2
networkx=2.5.1=pypi_0
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
oauth2client=3.0.0=pypi_0
oauthlib=3.1.0=py_0
opencv-python=4.5.2.52=pypi_0
openssl=1.1.1k=h2bbff1b_0
opt_einsum=3.1.0=py_0
pillow=8.2.0=pypi_0
pip=21.0.1=py36haa95532_0
playsound=1.2.2=pypi_0
protobuf=3.14.0=py36hd77b12b_1
psutil=5.8.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pycryptodome=3.4.3=pypi_0
pyfcm=1.5.1=pypi_0
pyjwt=1.7.1=py36_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pyreadline=2.1=py36_1
pyrebase=3.0.27=pypi_0
pysocks=1.7.1=py36haa95532_0
python=3.6.13=h3758d61_0
python-dateutil=2.8.1=pypi_0
python-jwt=2.0.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.0.3=pypi_0
requests=2.11.1=pypi_0
requests-oauthlib=1.3.0=py_0
requests-toolbelt=0.7.0=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.17.2=pypi_0
scipy=1.5.2=py36h9439919_0
setuptools=52.0.0=py36haa95532_0
six=1.15.0=py36haa95532_0
slidingwindow=0.0.14=pypi_0
sqlite=3.35.4=h2bbff1b_0
tabulate=0.8.9=pypi_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=1.15.0=gpu_py36h2b26d6b_0
tensorflow-base=1.15.0=gpu_py36h1afeea4_0
tensorflow-estimator=1.15.1=pyh2649769_0
tensorflow-gpu=1.15.0=h0d30ee6_0
tensorpack=0.11=pypi_0
termcolor=1.1.0=py36haa95532_1
tifffile=2020.9.3=pypi_0
torch=1.8.1+cu111=pypi_0
torchaudio=0.8.1=pypi_0
torchvision=0.9.1+cu111=pypi_0
tqdm=4.60.0=pypi_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=0.16.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py36haa95532_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.12.1=py36he774522_1
yarl=1.6.3=py36h2bbff1b_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
